requirements.txt
tqdm
pyyaml
numpy
opencv-contrib-python
matplotlib
